feat(user roles): add user roles feature
User roles are: admin, moderator, user. They have different permissions depending on role. This feature is necessary right now to prevent users from using the /users endpoints and get information about other users or modify other users

+++ go/db/moneyEntry.go
@@ -1,16 +1,22 @@
 package database
 
+import (
+	"errors"
+
+	"github.com/google/uuid"
+)
+
 type MoneyEntry struct {
-	ID        int64 `json:"id"`
-	Balance   float64 `json:"balance"`
-	Budget    float64 `json:"budget"`
-	Ratio     float64 `json:"ratio"`
-	CreatedAt string `json:"created_at"`
-	UserID    int64  `json:"user_id"`
+	ID        uuid.UUID `json:"id"`
+	Balance   float64   `json:"balance"`
+	Budget    float64   `json:"budget"`
+	Ratio     float64   `json:"ratio"`
+	CreatedAt string    `json:"created_at"`
+	UserID    uuid.UUID `json:"user_id"`
 }
 
-func (db *Database) InsertMoneyDB(entry *MoneyEntry) (int64, error) {
-	var id int64
+func (db *Database) InsertMoneyDB(entry *MoneyEntry) (uuid.UUID, error) {
+	var id uuid.UUID
 	err := db.DB.QueryRow(
 		"INSERT INTO money (balance, budget, ratio, user_id) VALUES ($1, $2, $3, $4) RETURNING id",
 		entry.Balance, entry.Budget, entry.Ratio, entry.UserID,
@@ -19,7 +25,10 @@ func (db *Database) InsertMoneyDB(entry *MoneyEntry) (int64, error) {
 	return id, err
 }
 
-func (db *Database) SelectUserMoneyDB(userID int64) ([]*MoneyEntry, error) {
+func (db *Database) SelectUserMoneyDB(userID *uuid.UUID) ([]*MoneyEntry, error) {
+	if userID == nil {
+		return nil, errors.New("userID is nil")
+	}
 	rows, err := db.DB.Query("SELECT id, balance, budget, ratio, created_at, user_id FROM money WHERE user_id = $1", userID)
 	if err != nil {
 		return nil, err
@@ -37,7 +46,10 @@ func (db *Database) SelectUserMoneyDB(userID int64) ([]*MoneyEntry, error) {
 	return entries, rows.Err()
 }
 
-func (db *Database) SelectUserMoneyByCountDB(userID int64, count int64) ([]*MoneyEntry, error) {
+func (db *Database) SelectUserMoneyByCountDB(userID *uuid.UUID, count int64) ([]*MoneyEntry, error) {
+	if userID == nil {
+		return nil, errors.New("userID is nil")
+	}
 	rows, err := db.DB.Query("SELECT id, balance, budget, ratio, created_at, user_id FROM money WHERE user_id = $1 ORDER BY created_at DESC LIMIT $2", userID, count)
 	if err != nil {
 		return nil, err
@@ -63,7 +75,10 @@ func (db *Database) UpdateMoneyDB(entry *MoneyEntry) error {
 	return err
 }
 
-func (db *Database) DeleteMoneyDB(id int64) error {
+func (db *Database) DeleteMoneyDB(id *uuid.UUID) error {
+	if id == nil {
+		return errors.New("id is nil")
+	}
 	_, err := db.DB.Exec(
 		"DELETE FROM money WHERE id = $1",
 		id,
